requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
pytz==2025.2
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
